requirements.txt
numpy
mtcnn
tensorflow-cpu
scipy
keras_vggface
keras_applications
streamlit
Pillow
